requirements.txt
osmnx>=2.0.0
networkx>=3.0
pandas>=2.0
geopandas>=0.14
numpy>=1.24
python-dotenv>=1.0
requests>=2.31
pytest>=7.0
matplotlib>=3.8
SciencePlots>=2.1
rich>=13.7
tqdm>=4.66
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic[email]>=2.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
httpx>=0.27
redis>=5.0
passlib>=1.7
python-multipart>=0.0.12
